Update statemachine. One initial state of patrol.
Statemachine is attached to the unit, this upon start up creates new patrol state.

Patrol state handles the patrolling of the unit. upon being within a certain distance of the player - the state should call an update to the statemanager with a new chase state added.
The chase state should handle the attack state - that is upon getting within a  closer range shots can be fired.

The attack state will then either revert back to chase state or return to patrol state upon loosing or destroying plyer unit