Updated attack state.
The bullet/missile has responsibility for the explosion  - since the missile causes the explosion then it has the responsibility of controlling the type of explosion.
Once added, the death or destruction of an game object will be controlled by the object its self and it will control the type of explosion required.

Need to update attackstate so that the tank does not get too close to other tank.