Added damage to AI Tank
Statemanager at the moment has the variable health - this and other such characteristic should be create as an object and injected into the statemanager for that tank.

Bullet controls the explosion of its damage - and provides damage to the statemanager setHealth.

The statemanager controls whether the tank is destroyed or not. Upon health reaching 0, vehicle explodes - controlled by statemanager..

Need to add damage to player vehicle.